requirements.txt
pyproj==1.9.6
pyasn1==0.4.5
pyasn1-modules==0.2.5
protobuf==3.7.1
Pillow==6.0.0
hl-etl-util==0.0.25
requests>=2.18.4
urllib3==1.22
PyYAML==3.13
simplejson>=3.16.0
pandas==0.24.2
pytz==2018.9
geopy>=1.19.0
shapely[vectorized]>=1.6.4.post2
geopandas>=0.4.1
swifter>=0.288